Changed all hangout pages to be more broad and added in functionailty to accomidate for events and hangouts.
