+++ client/components/app.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
-import HangoutsList from './hangouts-list';
+import ListPage from './list-page';
 import HangoutDetails from './hangout-details';
-import CreateHangout from './create-hangouts';
+import CreatePage from './create-page';
 import Sidebar from './sidebar';
 import {
   BrowserRouter as Router,
   Switch,
   Route
 } from 'react-router-dom';
+import FakeHomePage from './placeholder-homepage';
 
 export default class App extends React.Component {
   constructor(props) {
@@ -29,9 +30,12 @@ export default class App extends React.Component {
       <Router>
         <Sidebar toggleSidebar={this.toggleSidebar} isSidebarHidden={this.state.isSidebarHidden} />
         <Switch>
-          <Route exact path="/hangouts" render={props => <HangoutsList {...props} toggleSidebar={this.toggleSidebar} />} />
+          <Route exact path="/" render={props => <FakeHomePage {...props} toggleSidebar={this.toggleSidebar} />} />
+          <Route exact path="/hangouts" render={props => <ListPage {...props} toggleSidebar={this.toggleSidebar} />} />
+          <Route exact path="/events" render={props => <ListPage {...props} toggleSidebar={this.toggleSidebar} />} />
           <Route path="/hangouts/:id" component={HangoutDetails} />
-          <Route path="/create/hangout" component={CreateHangout} />
+          <Route path="/create/hangout" component={CreatePage} />
+          <Route path="/create/event" component={CreatePage} />
         </Switch>
       </Router>
     );
